smart-contract-verifier: rename 'middleware_types' to 'types'
